expense-tracker-web for roadmap sh website/requirements.txt
Flask
Flask-PyMongo
Flask-Login
Flask-Bcrypt
pandas
reportlab
requests
python-dotenv
pytesseract
Pillow
Flask-Mail 